refactor(lotto): Move the location of the function that converts from domain to dto to controller
