+++ src/main/java/domain/Lottos.java
@@ -20,27 +20,7 @@ public class Lottos {
         lottos.add(lotto);
     }
 
-    public LottosResponse getLottosDto() {
-        List<LottoResponse> lottoResponses = new ArrayList<>();
-        for (Lotto lotto : lottos) {
-            lottoResponses.add(lotto.getLottoDto());
-        }
-        return new LottosResponse(lottoResponses);
-    }
-
-    public ResultResponse getResult(WinningLotto winningLotto, Amount amount) {
-        EnumMap<Rank, Integer> countRank = countMatchNumbers(winningLotto);
-
-        long prizeSum = 0L;
-        for (Entry<Rank, Integer> rankIntegerEntry : countRank.entrySet()) {
-            prizeSum += Rank.calculateTotalPrize(rankIntegerEntry.getKey(), rankIntegerEntry.getValue());
-        }
-
-        double profit = amount.calculateProfit(prizeSum);
-        return new ResultResponse(countRank, profit);
-    }
-
-    private EnumMap<Rank, Integer> countMatchNumbers(WinningLotto winningLotto) {
+    public EnumMap<Rank, Integer> countMatchNumbers(WinningLotto winningLotto) {
         EnumMap<Rank, Integer> countRank = new EnumMap<>(Rank.class);
 
         for (Rank rank : Rank.values()) {
@@ -53,5 +33,9 @@ public class Lottos {
         }
 
         return countRank;
+    }
+
+    public List<Lotto> getLottos() {
+        return lottos;
     }
 }
